Test: Move more tests to dedicated modules
