+++ hevm.cabal
@@ -289,6 +289,7 @@ test-suite test
     EVM.Expr.ExprTests
     EVM.Expr.Generator
     EVM.ConcreteExecution.ConcreteExecutionTests
+    EVM.SymExec.SymExecTests
     EVM.Test.FuzzSymExec
   build-depends:
     base16,
